Combine block utilities
